+++ contracts/mock_adapter/src/tests/gains_losses.rs
@@ -1,7 +1,3 @@
-pub mod gains_losses;
-pub mod migration;
-pub mod tm_unbond;
-
 use shade_multi_test::interfaces::{
     dao::{
         init_dao,
@@ -23,301 +19,7 @@ use shade_protocol::{
     utils::cycle::Cycle,
 };
 
-pub fn dao_int_test(
-    initial_treasury_bal: Uint128,
-    snip20_symbol: &str,
-    allow_amount: Vec<Uint128>,
-    allow_type: Vec<AllowanceType>,
-    cycle: Vec<Cycle>,
-    allow_tolerance: Vec<Uint128>,
-    expected_allowance: Vec<Uint128>,
-    alloc_amount: Vec<Vec<Uint128>>,
-    alloc_type: Vec<Vec<AllocationType>>,
-    alloc_tolerance: Vec<Vec<Uint128>>,
-    is_instant_unbond: bool,
-    expected_treasury: Uint128,
-    expected_manager: Vec<Uint128>,
-    expected_adapter: Vec<Vec<Uint128>>,
-) {
-    let mut app = App::default();
-    let mut contracts = DeployedContracts::new();
-    let num_managers = allow_amount.len();
-    init_dao(
-        &mut app,
-        "admin",
-        &mut contracts,
-        initial_treasury_bal,
-        snip20_symbol.clone(),
-        allow_type.clone(),
-        cycle.clone(),
-        allow_amount.clone(),
-        allow_tolerance.clone(),
-        alloc_type.clone(),
-        alloc_amount.clone(),
-        alloc_tolerance.clone(),
-        is_instant_unbond,
-    );
-    //query assets
-    let assets_query_res = treasury::assets_query(&app, &contracts).unwrap();
-    assert!(
-        assets_query_res.contains(
-            &contracts
-                .get(&SupportedContracts::Snip20(snip20_symbol.to_string()))
-                .unwrap()
-                .address
-        )
-    );
-    //query allowance
-    for i in 0..num_managers {
-        assert_eq!(
-            expected_allowance[i],
-            treasury::allowance_query(
-                &app,
-                "admin",
-                &contracts,
-                snip20_symbol.to_string(),
-                SupportedContracts::TreasuryManager(i)
-            )
-            .unwrap(),
-            "Treasury->Manager Allowance",
-        );
-    }
-    let mut bals;
-    if is_instant_unbond {
-        bals = system_balance_reserves(&app, &contracts, snip20_symbol.to_string());
-    } else {
-        bals = system_balance_unbondable(&app, &contracts, snip20_symbol.to_string());
-    }
-    println!("{:?}", bals);
-    assert_eq!(bals.0, expected_treasury);
-    for (i, manager_tuples) in bals.1.iter().enumerate() {
-        assert_eq!(manager_tuples.0, expected_manager[i]);
-        for (j, adapter_bals) in manager_tuples.1.iter().enumerate() {
-            assert_eq!(adapter_bals.clone(), expected_adapter[i][j]);
-        }
-    }
-    let mut k = 0;
-    for i in 0..num_managers {
-        treasury::allowance(
-            &mut app,
-            "admin",
-            &contracts,
-            snip20_symbol.to_string(),
-            i,
-            allow_type[i].clone(),
-            cycle[i].clone(),
-            Uint128::zero(),
-            allow_tolerance[i].clone(),
-        );
-        for j in 0..alloc_amount[i].len() {
-            treasury_manager::allocate(
-                &mut app,
-                "admin",
-                &contracts,
-                snip20_symbol.to_string(),
-                Some(j.to_string()),
-                &SupportedContracts::MockAdapter(k),
-                alloc_type[i][j].clone(),
-                Uint128::zero(),
-                alloc_tolerance[i][j].clone(),
-                i,
-            );
-            k += 1;
-        }
-        k += 1;
-    }
-
-    update_dao(&mut app, "admin", &contracts, "SSCRT", num_managers);
-    treasury::update_exec(&mut app, "admin", &contracts, "SSCRT".to_string());
-    println!("{:?}", bals);
-    if !is_instant_unbond {
-        k = 0;
-        for i in 0..num_managers {
-            for j in 0..alloc_amount[i].len() {
-                println!("{}", k);
-                mock_adapter_complete_unbonding(
-                    &mut app,
-                    "admin",
-                    &contracts,
-                    SupportedContracts::MockAdapter(k),
-                )
-                .unwrap();
-                k += 1;
-            }
-            k += 1;
-        }
-        update_dao(&mut app, "admin", &contracts, "SSCRT", num_managers);
-        bals = system_balance_unbondable(&app, &contracts, "SSCRT".to_string());
-    } else {
-        bals = system_balance_reserves(&app, &contracts, "SSCRT".to_string());
-    }
-    assert_eq!(bals.0, initial_treasury_bal);
-    for (i, manager_tuples) in bals.1.iter().enumerate() {
-        assert_eq!(manager_tuples.0, Uint128::zero());
-        for (j, adapter_bals) in manager_tuples.1.iter().enumerate() {
-            assert_eq!(adapter_bals.clone(), Uint128::zero());
-        }
-    }
-}
-
-macro_rules! dao_tests {
-    ($($name:ident: $value:expr,)*) => {
-        $(
-            #[test]
-            fn $name() {
-                let (
-                    initial_treasury_bal,
-                    snip20_symbol,
-                    allow_amount,
-                    allow_type,
-                    cycle,
-                    allow_tolerance,
-                    expected_allowance,
-                    alloc_amount,
-                    alloc_type,
-                    alloc_tolerance,
-                    is_instant_unbond,
-                    expected_treasury,
-                    expected_manager,
-                    expected_adapter,
-                ) = $value;
-                dao_int_test(
-                    initial_treasury_bal,
-                    snip20_symbol,
-                    allow_amount,
-                    allow_type,
-                    cycle,
-                    allow_tolerance,
-                    expected_allowance,
-                    alloc_amount,
-                    alloc_type,
-                    alloc_tolerance,
-                    is_instant_unbond,
-                    expected_treasury,
-                    expected_manager,
-                    expected_adapter,
-                );
-            }
-        )*
-    }
-}
-
-dao_tests! {
-    dao_test_0:(
-        Uint128::new(1_000_000),
-        "SSCRT",
-        vec![Uint128::new(5 * 10u128.pow(17))],
-        vec![AllowanceType::Portion],
-        vec![Cycle::Constant],
-        vec![Uint128::zero()],
-        vec![Uint128::new(90_000)],
-        vec![vec![Uint128::new(1 * 10u128.pow(17)), Uint128::new(400_000)]],
-        vec![vec![AllocationType::Portion, AllocationType::Amount]],
-        vec![vec![Uint128::zero(), Uint128::zero()]],
-        true,
-        Uint128::new(590_000),
-        vec![Uint128::new(0)],
-        vec![vec![Uint128::new(10_000), Uint128::new(400_000)]],
-    ),
-    dao_test_1:(
-        Uint128::new(100_000_000),
-        "SSCRT",
-        vec![Uint128::new(50_000_000), Uint128::new(40_000_000)],
-        vec![AllowanceType::Amount, AllowanceType::Amount],
-        vec![Cycle::Constant, Cycle::Constant],
-        vec![Uint128::zero(), Uint128::zero()],
-        vec![Uint128::new(21_000_000), Uint128::new(18_000_000)],
-        vec![vec![Uint128::new(5 * 10u128.pow(17)), Uint128::new(4_000_000), Uint128::new(4_000_000)], vec![Uint128::new(5 * 10u128.pow(17)), Uint128::new(4_000_000)]],
-        vec![vec![AllocationType::Portion, AllocationType::Amount, AllocationType::Amount],vec![AllocationType::Portion, AllocationType::Amount]],
-        vec![vec![Uint128::zero(), Uint128::zero(), Uint128::zero()],vec![Uint128::zero(), Uint128::zero()]],
-        true,
-        Uint128::new(49_000_000),
-        vec![Uint128::new(0), Uint128::new(0)],
-        vec![vec![Uint128::new(21_000_000), Uint128::new(4_000_000), Uint128::new(4_000_000)],vec![Uint128::new(18_000_000), Uint128::new(4_000_000)]],
-    ),
-    dao_test_2:(
-        Uint128::new(100),
-        "SSCRT",
-        vec![Uint128::new(5 * 10u128.pow(17))],
-        vec![AllowanceType::Portion],
-        vec![Cycle::Constant],
-        vec![Uint128::zero()],
-        vec![Uint128::new(9)],
-        vec![vec![Uint128::new(1 * 10u128.pow(17)), Uint128::new(40)]],
-        vec![vec![AllocationType::Portion, AllocationType::Amount]],
-        vec![vec![Uint128::zero(), Uint128::zero()]],
-        true,
-        Uint128::new(59),
-        vec![Uint128::new(0)],
-        vec![vec![Uint128::new(1), Uint128::new(40)]],
-    ),
-    dao_test_3: (
-        Uint128::new(1000),
-        "SSCRT",
-        vec![
-            Uint128::new(50), // Amount - 50
-            Uint128::new(6 * 10u128.pow(17)), // Poriton - 60%
-            Uint128::new(100), // Amount - 100
-            Uint128::new(4 * 10u128.pow(17)), // Portion - 40%
-        ], // Allowance amount
-        vec![AllowanceType::Amount, AllowanceType::Portion, AllowanceType::Amount,  AllowanceType::Portion],
-        vec![Cycle::Constant; 4],
-        vec![Uint128::zero(); 4],
-        vec![Uint128::new(6), Uint128::new(98), Uint128::new(16),  Uint128::new(64)],
-        vec![
-            vec![Uint128::new(6 * 10u128.pow(17)), Uint128::new(5), Uint128::new(2 * 10u128.pow(17)), Uint128::new(15)];4
-        ],
-        vec![
-            vec![AllocationType::Portion, AllocationType::Amount, AllocationType::Portion, AllocationType::Amount];4
-        ],
-        vec![
-            vec![Uint128::zero(); 4]; 4
-        ],
-        true,
-        Uint128::new(184),
-        vec![Uint128::zero(); 4],
-        vec![
-            vec![Uint128::new(18), Uint128::new(5), Uint128::new(6), Uint128::new(15)],
-            vec![Uint128::new(294), Uint128::new(5), Uint128::new(98), Uint128::new(15)],
-            vec![Uint128::new(48), Uint128::new(5), Uint128::new(16), Uint128::new(15)],
-            vec![Uint128::new(192), Uint128::new(5), Uint128::new(64), Uint128::new(15)],
-        ]
-    ),
-    dao_test_adapter_unbonding: (
-        Uint128::new(1000),
-        "SSCRT",
-        vec![
-            Uint128::new(50), // Amount - 50
-            Uint128::new(6 * 10u128.pow(17)), // Poriton - 60%
-            Uint128::new(100), // Amount - 100
-            Uint128::new(4 * 10u128.pow(17)), // Portion - 40%
-        ], // Allowance amount
-        vec![AllowanceType::Amount, AllowanceType::Portion, AllowanceType::Amount,  AllowanceType::Portion],
-        vec![Cycle::Constant; 4],
-        vec![Uint128::zero(); 4],
-        vec![Uint128::new(6), Uint128::new(98), Uint128::new(16),  Uint128::new(64)],
-        vec![
-            vec![Uint128::new(6 * 10u128.pow(17)), Uint128::new(5), Uint128::new(2 * 10u128.pow(17)), Uint128::new(15)];4
-        ],
-        vec![
-            vec![AllocationType::Portion, AllocationType::Amount, AllocationType::Portion, AllocationType::Amount];4
-        ],
-        vec![
-            vec![Uint128::zero(); 4]; 4
-        ],
-        false,
-        Uint128::new(184),
-        vec![Uint128::zero(); 4],
-        vec![
-            vec![Uint128::new(18), Uint128::new(5), Uint128::new(6), Uint128::new(15)],
-            vec![Uint128::new(294), Uint128::new(5), Uint128::new(98), Uint128::new(15)],
-            vec![Uint128::new(48), Uint128::new(5), Uint128::new(16), Uint128::new(15)],
-            vec![Uint128::new(192), Uint128::new(5), Uint128::new(64), Uint128::new(15)],
-        ]
-    ),
-}
-
-/*pub fn dao_int_gains_losses(
+pub fn dao_int_gains_losses(
     initial_treasury_bal: Uint128,
     allow_type: Vec<AllowanceType>,
     t_cycle: Vec<Cycle>,
@@ -1285,326 +987,3 @@ dao_tests_gains_losses! {
         ]),
     ),
 }
-
-pub fn test_tm_unbond(
-    unbond_amount: Uint128,
-    adapter_gain_amount: Option<Uint128>,
-    amount_adapter_bal: (Uint128, Uint128),
-    expected_before_unbond: (Uint128, Vec<(Uint128, Vec<Uint128>)>),
-    expected_after_unbond: (Uint128, Vec<(Uint128, Vec<Uint128>)>),
-) {
-    let mut app = App::default();
-    let mut contracts = DeployedContracts::new();
-    init_dao(
-        &mut app,
-        "admin",
-        &mut contracts,
-        Uint128::new(1000),
-        "SSCRT",
-        vec![AllowanceType::Amount],
-        vec![Cycle::Constant],
-        vec![
-            Uint128::new(500), // Amount - 500
-        ], // Allowance amount
-        vec![Uint128::zero()],
-        vec![
-            vec![
-                AllocationType::Portion,
-                AllocationType::Amount,
-                AllocationType::Portion,
-                AllocationType::Amount
-            ];
-            4
-        ],
-        vec![
-            vec![
-                Uint128::new(6 * 10u128.pow(17)),
-                amount_adapter_bal.0,
-                Uint128::new(2 * 10u128.pow(17)),
-                amount_adapter_bal.1,
-            ];
-            4
-        ],
-        vec![vec![Uint128::zero(); 4]; 4],
-        true,
-    );
-    let bals = system_balance_reserves(&app, &contracts, "SSCRT".to_string());
-    assert_eq!(bals, expected_before_unbond);
-    match adapter_gain_amount {
-        Some(x) => {
-            for i in vec![1, 3] {
-                snip20::send(
-                    &mut app,
-                    "admin",
-                    &contracts,
-                    "SSCRT".to_string(),
-                    contracts
-                        .get(&SupportedContracts::MockAdapter(i))
-                        .unwrap()
-                        .address
-                        .to_string(),
-                    x,
-                    None,
-                );
-            }
-        }
-        None => {}
-    }
-    treasury_manager::unbond_exec(
-        &mut app,
-        "admin",
-        &contracts,
-        "SSCRT".to_string(),
-        SupportedContracts::TreasuryManager(0),
-        unbond_amount,
-    )
-    .unwrap();
-
-    let bals = system_balance_reserves(&app, &contracts, "SSCRT".to_string());
-    assert_eq!(bals, expected_after_unbond);
-}
-
-macro_rules! dao_tests_tm_unbond {
-    ($($name:ident: $value:expr,)*) => {
-        $(
-            #[test]
-            fn $name() {
-                let (
-                    unbond_amount,
-                    adapter_gain_amount,
-                    amount_adapter_bal,
-                    expected_before_unbond,
-                    expected_after_unbond,
-                ) = $value;
-                test_tm_unbond(
-                    unbond_amount,
-                    adapter_gain_amount,
-                    amount_adapter_bal,
-                    expected_before_unbond,
-                    expected_after_unbond,
-                );
-            }
-        )*
-    }
-}
-
-dao_tests_tm_unbond! {
-    unbond_only_from_amount_adapters:(
-        Uint128::new(7),
-        Some(Uint128::new(10)),
-        (Uint128::new(10), Uint128::new(20)),
-        (Uint128::new(594), vec![(Uint128::new(0), vec![
-            Uint128::new(282),
-            Uint128::new(10),
-            Uint128::new(94),
-            Uint128::new(20),
-        ])]),
-        (Uint128::new(594), vec![(Uint128::new(7), vec![
-            Uint128::new(282),
-            Uint128::new(17),
-            Uint128::new(94),
-            Uint128::new(26)
-        ])])
-    ),
-    unbond_case_extra_from_amount_adapters_and_some_from_portion_adapters:(
-        Uint128::new(21),
-        Some(Uint128::new(10)),
-        (Uint128::new(10), Uint128::new(20)),
-        (Uint128::new(594), vec![(Uint128::new(0), vec![
-            Uint128::new(282),
-            Uint128::new(10),
-            Uint128::new(94),
-            Uint128::new(20),
-        ])]),
-        (Uint128::new(594), vec![(Uint128::new(21), vec![
-            Uint128::new(282),
-            Uint128::new(10),
-            Uint128::new(93),
-            Uint128::new(20)
-        ])])
-    ),
-    unbond_case_extra_from_amount_adapters_and_all_from_portion_adapters:(
-        Uint128::new(396),
-        Some(Uint128::new(10)),
-        (Uint128::new(10), Uint128::new(20)),
-        (Uint128::new(594), vec![(Uint128::new(0), vec![
-            Uint128::new(282),
-            Uint128::new(10),
-            Uint128::new(94),
-            Uint128::new(20),
-        ])]),
-        (Uint128::new(594), vec![(Uint128::new(396), vec![
-            Uint128::new(0),
-            Uint128::new(10),
-            Uint128::new(0),
-            Uint128::new(20)
-        ])])
-    ),
-    unbond_case_extra_and_some_from_amount_adapters_and_all_from_portion_adapters:(
-        Uint128::new(391),
-        None,
-        (Uint128::new(100), Uint128::new(200)),
-        (Uint128::new(540), vec![(Uint128::new(0), vec![
-            Uint128::new(120),
-            Uint128::new(100),
-            Uint128::new(40),
-            Uint128::new(200),
-        ])]),
-        (Uint128::new(540), vec![(Uint128::new(391), vec![
-            Uint128::new(0),
-            Uint128::new(23),
-            Uint128::new(0),
-            Uint128::new(46)
-        ])])
-    ),
-    unbond_all:(
-        Uint128::new(460),
-        None,
-        (Uint128::new(100), Uint128::new(200)),
-        (Uint128::new(540), vec![(Uint128::new(0), vec![
-            Uint128::new(120),
-            Uint128::new(100),
-            Uint128::new(40),
-            Uint128::new(200),
-        ])]),
-        (Uint128::new(540), vec![(Uint128::new(460), vec![
-            Uint128::new(0),
-            Uint128::new(0),
-            Uint128::new(0),
-            Uint128::new(0)
-        ])])
-    ),
-}
-
-pub fn migration_test(is_instant_unbond: bool) {
-    const multisig: &str = "multisig";
-    let mut app = App::default();
-    let mut contracts = DeployedContracts::new();
-    init_dao(
-        &mut app,
-        "admin",
-        &mut contracts,
-        Uint128::new(1500),
-        "SSCRT",
-        vec![
-            AllowanceType::Amount,
-            AllowanceType::Portion,
-            AllowanceType::Amount,
-            AllowanceType::Portion,
-        ],
-        vec![Cycle::Constant; 4],
-        vec![
-            Uint128::new(200),                // Amount - 50
-            Uint128::new(6 * 10u128.pow(17)), // Poriton - 60%
-            Uint128::new(300),                // Amount - 100
-            Uint128::new(3 * 10u128.pow(17)), // Portion - 40%
-        ], // Allowance amount
-        vec![Uint128::zero(); 4],
-        vec![
-            vec![
-                AllocationType::Portion,
-                AllocationType::Amount,
-                AllocationType::Portion,
-                AllocationType::Amount
-            ];
-            4
-        ],
-        vec![
-            vec![
-                Uint128::new(6 * 10u128.pow(17)),
-                Uint128::new(50),
-                Uint128::new(2 * 10u128.pow(17)),
-                Uint128::new(75),
-            ];
-            4
-        ],
-        vec![vec![Uint128::zero(); 4]; 4],
-        true,
-    );
-    snip20::set_viewing_key(
-        &mut app,
-        multisig,
-        &contracts,
-        "SSCRT".to_string(),
-        multisig.to_string(),
-    )
-    .unwrap();
-    treasury::set_config(&mut app, "admin", &contracts, dao::treasury::Config {
-        admin_auth: contracts
-            .get(&SupportedContracts::AdminAuth)
-            .unwrap()
-            .clone()
-            .into(),
-        multisig: Addr::unchecked(multisig),
-    })
-    .unwrap();
-    treasury::set_run_level(
-        &mut app,
-        "admin",
-        &contracts,
-        dao::treasury::RunLevel::Migrating,
-    )
-    .unwrap();
-    update_dao(&mut app, "admin", &contracts, "SSCRT", 4);
-    if is_instant_unbond {
-        update_dao(&mut app, "admin", &contracts, "SSCRT", 4);
-    } else {
-        let mut k = 0;
-        for i in 0..4 {
-            for j in 0..4 {
-                mock_adapter_complete_unbonding(
-                    &mut app,
-                    "admin",
-                    &contracts,
-                    SupportedContracts::MockAdapter(k),
-                )
-                .unwrap();
-                k += 1;
-            }
-            k += 1;
-        }
-        update_dao(&mut app, "admin", &contracts, "SSCRT", 4);
-        update_dao(&mut app, "admin", &contracts, "SSCRT", 4);
-    }
-    println!(
-        "{:?}\n{:?}",
-        system_balance_reserves(&app, &contracts, "SSCRT".to_string()),
-        system_balance_unbondable(&app, &contracts, "SSCRT".to_string())
-    );
-    assert_eq!(
-        snip20::balance_query(
-            &app,
-            multisig,
-            &contracts,
-            "SSCRT".to_string(),
-            multisig.to_string()
-        )
-        .unwrap(),
-        Uint128::new(1500)
-    );
-}
-
-macro_rules! dao_tests_migration {
-    ($($name:ident: $value:expr,)*) => {
-        $(
-            #[test]
-            fn $name() {
-                let (
-                    is_instant_unbond,
-                ) = $value;
-                migration_test(
-                    is_instant_unbond,
-                );
-            }
-        )*
-    }
-}
-
-dao_tests_migration! (
-    dao_test_migration_instant_unbond: (
-        true,
-    ),
-    dao_test_migration_non_instant_unbond: (
-        false,
-    ),
-);*/
